# openapi.json
{"openapi":"3.0.1","info":{"title":"openapi","version":"3.0.1","x-logo":{"url":"https://fastapi.tiangolo.com/img/logo-margin/logo-teal.png"}},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/quests":{"get":{"summary":"Get Quests","operationId":"get_quests_quests_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/DetailedQuest"},"type":"array","title":"Response Get Quests Quests Get"}}}}}},"post":{"summary":"Create Quest","operationId":"create_quest_quests_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateQuest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/quests/{quest_id}":{"patch":{"summary":"Edit Quest","operationId":"edit_quest_quests__quest_id__patch","parameters":[{"name":"quest_id","in":"path","required":true,"schema":{"title":"Quest Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Quest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/quests/{quest_id}/answer":{"post":{"summary":"Answer Brain","operationId":"answer_brain_quests__quest_id__answer_post","parameters":[{"name":"quest_id","in":"path","required":true,"schema":{"title":"Quest Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BrainAnswer"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/quests/{quest_id}/participate":{"post":{"summary":"Participiate Brawn","operationId":"participiate_brawn_quests__quest_id__participate_post","parameters":[{"name":"quest_id","in":"path","required":true,"schema":{"title":"Quest Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BrawnParticipant"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/quests/{quest_id}/complete":{"post":{"summary":"Complete Quest","operationId":"complete_quest_quests__quest_id__complete_post","parameters":[{"name":"quest_id","in":"path","required":true,"schema":{"title":"Quest Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/quests/{quest_id}/pending":{"post":{"summary":"Pending Quest","operationId":"pending_quest_quests__quest_id__pending_post","parameters":[{"name":"quest_id","in":"path","required":true,"schema":{"title":"Quest Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/quests/{quest_id}/cancel":{"post":{"summary":"Cancel Quest","operationId":"cancel_quest_quests__quest_id__cancel_post","parameters":[{"name":"quest_id","in":"path","required":true,"schema":{"title":"Quest Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users":{"get":{"summary":"Get Users","operationId":"get_users_users_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"$ref":"#/components/schemas/User"},"type":"object","title":"Response Get Users Users Get"}}}}}}},"/answer/{quest_id}/toggle":{"post":{"summary":"Toggle Favorite","operationId":"toggle_favorite_answer__quest_id__toggle_post","parameters":[{"name":"quest_id","in":"path","required":true,"schema":{"title":"Quest Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","title":"Json"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BrainAnswer":{"properties":{"quest":{"type":"integer","title":"Quest"},"author":{"type":"integer","title":"Author"},"text":{"type":"string","title":"Text"},"accepted":{"type":"boolean","title":"Accepted"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["quest","author","text","accepted","created_at"],"title":"BrainAnswer"},"BrawnParticipant":{"properties":{"quest":{"type":"integer","title":"Quest"},"user":{"type":"integer","title":"User"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["quest","user","created_at"],"title":"BrawnParticipant"},"CreateQuest":{"properties":{"type":{"$ref":"#/components/schemas/QuestType"},"author":{"type":"integer","title":"Author"},"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"status":{"$ref":"#/components/schemas/Status"},"trait":{"$ref":"#/components/schemas/Trait"},"deadline":{"type":"string","format":"date-time","title":"Deadline"},"max_participants":{"type":"integer","title":"Max Participants"}},"type":"object","required":["type","author","title","description","status","trait","deadline","max_participants"],"title":"CreateQuest"},"DetailedQuest":{"properties":{"id":{"type":"integer","title":"Id"},"type":{"$ref":"#/components/schemas/QuestType"},"author":{"type":"integer","title":"Author"},"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"status":{"$ref":"#/components/schemas/Status"},"trait":{"$ref":"#/components/schemas/Trait"},"deadline":{"type":"string","format":"date-time","title":"Deadline"},"max_participants":{"type":"integer","title":"Max Participants"},"brain_answers":{"items":{"$ref":"#/components/schemas/BrainAnswer"},"type":"array","title":"Brain Answers"},"brawn_participants":{"items":{"$ref":"#/components/schemas/BrawnParticipant"},"type":"array","title":"Brawn Participants"}},"type":"object","required":["id","type","author","title","description","status","trait","deadline","max_participants","brain_answers","brawn_participants"],"title":"DetailedQuest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Quest":{"properties":{"id":{"type":"integer","title":"Id"},"type":{"$ref":"#/components/schemas/QuestType"},"author":{"type":"integer","title":"Author"},"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"status":{"$ref":"#/components/schemas/Status"},"trait":{"$ref":"#/components/schemas/Trait"},"deadline":{"type":"string","format":"date-time","title":"Deadline"},"max_participants":{"type":"integer","title":"Max Participants"}},"type":"object","required":["id","type","author","title","description","status","trait","deadline","max_participants"],"title":"Quest"},"QuestType":{"type":"integer","enum":[1,2],"title":"QuestType"},"Status":{"type":"integer","enum":[1,2,3,4],"title":"Status"},"Trait":{"type":"integer","enum":[1,2,3,4],"title":"Trait"},"User":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"str":{"type":"integer","title":"Str"},"dex":{"type":"integer","title":"Dex"},"cha":{"type":"integer","title":"Cha"},"int":{"type":"integer","title":"Int"}},"type":"object","required":["id","name","str","dex","cha","int"],"title":"User"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}